Fix location seed normalization
