+++ apps/web/lib/location-resolution/normalize.ts
@@ -1,13 +1,4 @@
-export function normalizeLocalityName(value: string): string {
-  return value
-    .normalize("NFD")
-    .replace(/[̀-ͯ]/g, "")
-    .normalize("NFC")
-    .toLowerCase()
-    .trim()
-    .replace(/\s+/g, " ");
-}
-
-export function namesAreExactNormalizedMatch(left: string, right: string): boolean {
-  return normalizeLocalityName(left) === normalizeLocalityName(right);
-}
+export {
+  namesAreExactNormalizedMatch,
+  normalizeLocalityName,
+} from "@dpf/db/location-normalize";
